test: 收敛 Respond Todo Tool Loop 测试
