+++ qq-maid-core/src/runtime/respond/tests/chat.rs
@@ -188,102 +188,6 @@ async fn private_strong_todo_reference_without_context_enters_tool_loop_and_clar
 }
 
 #[tokio::test]
-async fn private_tool_loop_registers_todo_tools_and_keeps_internal_ids_hidden() {
-    let inspector = MockProvider::new().with_tool_protocol(ToolCallingProtocol::OpenAiResponses);
-    let service = test_service_with_provider_and_tool_calling(inspector.clone(), true);
-    let owner = TodoStore::owner(Some("u1"), "private:u1");
-    service
-        .todo_store
-        .create(
-            &owner,
-            TodoItemDraft {
-                title: "检查机器人日志".to_owned(),
-                detail: None,
-                raw_text: None,
-                due_date: None,
-                due_at: None,
-                reminder_at: None,
-                time_precision: TodoTimePrecision::None,
-                recurrence_kind: crate::runtime::todo::TodoRecurrenceKind::None,
-                recurrence_interval_days: 0,
-                recurrence_interval: 0,
-                recurrence_unit: crate::runtime::todo::TodoRecurrenceUnit::Day,
-            },
-        )
-        .unwrap();
-
-    service
-        .respond(private_message("杭州今天要带伞吗"))
-        .await
-        .unwrap();
-    let tool_request = inspector.tool_requests().remove(0);
-    let list_output = tool_request
-        .tools
-        .execute_json(
-            &tool_request.tool_context,
-            "list_todos",
-            r#"{"status":"pending"}"#,
-        )
-        .await
-        .unwrap();
-    let listed: Value = serde_json::from_str(&list_output).unwrap();
-    assert_eq!(listed["items"][0]["visible_number"], 1);
-    assert!(listed["items"][0].get("id").is_none());
-
-    let complete_output = tool_request
-        .tools
-        .execute_json(
-            &tool_request.tool_context,
-            "complete_todos",
-            r#"{"numbers":[1]}"#,
-        )
-        .await
-        .unwrap();
-    let completed: Value = serde_json::from_str(&complete_output).unwrap();
-    assert_eq!(completed["completed"][0]["title"], "检查机器人日志");
-    assert!(completed["completed"][0].get("id").is_none());
-    assert_eq!(
-        service.todo_store.list_all(&owner).unwrap()[0].status,
-        TodoStatus::Completed
-    );
-    let list_completed_output = tool_request
-        .tools
-        .execute_json(
-            &tool_request.tool_context,
-            "list_todos",
-            r#"{"status":"completed"}"#,
-        )
-        .await
-        .unwrap();
-    let listed_completed: Value = serde_json::from_str(&list_completed_output).unwrap();
-    assert_eq!(listed_completed["items"][0]["visible_number"], 1);
-    let restore_output = tool_request
-        .tools
-        .execute_json(
-            &tool_request.tool_context,
-            "restore_todos",
-            r#"{"numbers":[1]}"#,
-        )
-        .await
-        .unwrap();
-    let restored: Value = serde_json::from_str(&restore_output).unwrap();
-    assert_eq!(restored["ok"], true);
-    assert_eq!(restored["restored"][0]["visible_number"], 1);
-    assert!(restored["missing_numbers"].as_array().unwrap().is_empty());
-
-    let session = service
-        .session_store
-        .get_or_create_active(&private_test_meta())
-        .unwrap();
-    assert!(session.last_todo_query.is_none());
-    let last_action = session.last_todo_action.expect("missing last_todo_action");
-    assert_eq!(last_action.owner_key, owner.key);
-    assert_eq!(last_action.title, "检查机器人日志");
-    assert_eq!(last_action.action, "restored");
-    assert_eq!(last_action.resulting_status, TodoStatus::Pending);
-}
-
-#[tokio::test]
 async fn private_tool_loop_can_query_train_schedule_with_trusted_rendering() {
     let inspector = MockProvider::new()
         .with_tool_protocol(ToolCallingProtocol::OpenAiResponses)
@@ -577,109 +481,6 @@ async fn group_tool_loop_exposes_query_only_tools_when_enabled() {
         diagnostics["tool_loop_enabled_tools"],
         serde_json::json!(["get_weather", "get_train_schedule", "get_rss_recent_items"])
     );
-}
-
-#[tokio::test]
-async fn todo_tools_create_cancel_restore_and_delete_use_existing_pending_boundaries() {
-    let inspector = MockProvider::new().with_tool_protocol(ToolCallingProtocol::OpenAiResponses);
-    let service = test_service_with_provider_and_tool_calling(inspector.clone(), true);
-    let owner = TodoStore::owner(Some("u1"), "private:u1");
-
-    service
-        .respond(private_message("帮我记待办"))
-        .await
-        .unwrap();
-    let tool_request = inspector.tool_requests().remove(0);
-    let create_output = tool_request
-        .tools
-        .execute_json(
-            &tool_request.tool_context,
-            "create_todo",
-            r#"{"content":"今晚检查机器人日志","title":null,"detail":null,"due_date":null,"due_at":null,"time_precision":null}"#,
-        )
-        .await
-        .unwrap();
-    let created: Value = serde_json::from_str(&create_output).unwrap();
-    assert_eq!(created["ok"], true);
-    assert_eq!(created["created"]["title"], "今晚检查机器人日志");
-    assert!(created.get("requires_confirmation").is_none());
-    assert_eq!(service.todo_store.list_pending(&owner).unwrap().len(), 1);
-
-    tool_request
-        .tools
-        .execute_json(
-            &tool_request.tool_context,
-            "list_todos",
-            r#"{"status":"pending"}"#,
-        )
-        .await
-        .unwrap();
-    let cancel_output = tool_request
-        .tools
-        .execute_json(&tool_request.tool_context, "cancel_todo", r#"{"number":1}"#)
-        .await
-        .unwrap();
-    let cancel: Value = serde_json::from_str(&cancel_output).unwrap();
-    assert_eq!(cancel["ok"], true);
-    assert_eq!(cancel["cancelled"][0]["visible_number"], 1);
-    assert!(cancel["missing_numbers"].as_array().unwrap().is_empty());
-    assert_eq!(
-        service.todo_store.list_all(&owner).unwrap()[0].status,
-        TodoStatus::Cancelled
-    );
-
-    tool_request
-        .tools
-        .execute_json(
-            &tool_request.tool_context,
-            "list_todos",
-            r#"{"status":"cancelled"}"#,
-        )
-        .await
-        .unwrap();
-    let restore_output = tool_request
-        .tools
-        .execute_json(
-            &tool_request.tool_context,
-            "restore_todos",
-            r#"{"numbers":[1]}"#,
-        )
-        .await
-        .unwrap();
-    let restore: Value = serde_json::from_str(&restore_output).unwrap();
-    assert_eq!(restore["restored"][0]["visible_number"], 1);
-    assert!(restore["missing_numbers"].as_array().unwrap().is_empty());
-    let restored = service.todo_store.list_pending(&owner).unwrap();
-    assert_eq!(restored.len(), 1);
-    assert!(restored[0].cancelled_at.is_none());
-
-    service
-        .todo_store
-        .complete(&owner, &restored[0].id)
-        .unwrap();
-    tool_request
-        .tools
-        .execute_json(
-            &tool_request.tool_context,
-            "list_todos",
-            r#"{"status":"completed"}"#,
-        )
-        .await
-        .unwrap();
-    let delete_output = tool_request
-        .tools
-        .execute_json(
-            &tool_request.tool_context,
-            "delete_todos",
-            r#"{"numbers":[1]}"#,
-        )
-        .await
-        .unwrap();
-    let delete: Value = serde_json::from_str(&delete_output).unwrap();
-    assert_eq!(delete["requires_confirmation"], true);
-    assert_eq!(delete["pending_action"], "delete");
-    service.respond(private_message("确认")).await.unwrap();
-    assert!(service.todo_store.list_all(&owner).unwrap().is_empty());
 }
 
 #[tokio::test]
@@ -3406,657 +3207,6 @@ async fn todo_write_or_question_phrases_do_not_enter_natural_query_path() {
 }
 
 #[tokio::test]
-async fn deterministic_pending_query_then_tool_loop_complete_first_uses_latest_snapshot() {
-    let inspector = MockProvider::new().with_tool_protocol(ToolCallingProtocol::OpenAiResponses);
-    let service = test_service_with_provider_and_tool_calling(inspector.clone(), true);
-    let owner = TodoStore::owner(Some("u1"), "private:u1");
-    let first = service
-        .todo_store
-        .create(
-            &owner,
-            TodoItemDraft {
-                title: "测试代办".to_owned(),
-                detail: None,
-                raw_text: None,
-                due_date: None,
-                due_at: None,
-                reminder_at: None,
-                time_precision: TodoTimePrecision::None,
-                recurrence_kind: crate::runtime::todo::TodoRecurrenceKind::None,
-                recurrence_interval_days: 0,
-                recurrence_interval: 0,
-                recurrence_unit: crate::runtime::todo::TodoRecurrenceUnit::Day,
-            },
-        )
-        .unwrap();
-    let second = service
-        .todo_store
-        .create(
-            &owner,
-            TodoItemDraft {
-                title: "明天晚上搬到16栋".to_owned(),
-                detail: None,
-                raw_text: None,
-                due_date: None,
-                due_at: None,
-                reminder_at: None,
-                time_precision: TodoTimePrecision::None,
-                recurrence_kind: crate::runtime::todo::TodoRecurrenceKind::None,
-                recurrence_interval_days: 0,
-                recurrence_interval: 0,
-                recurrence_unit: crate::runtime::todo::TodoRecurrenceUnit::Day,
-            },
-        )
-        .unwrap();
-
-    let listed = service
-        .respond(private_message("看一下待办"))
-        .await
-        .unwrap();
-    assert_eq!(listed.command.as_deref(), Some("todo_list"));
-    let listed_text = listed.text.unwrap();
-    assert!(listed_text.contains("1. 测试代办"));
-    assert!(listed_text.contains("2. 明天晚上搬到16栋"));
-    assert_eq!(inspector.tool_call_count(), 0);
-
-    let session = service
-        .session_store
-        .get_or_create_active(&private_test_meta())
-        .unwrap();
-    let snapshot = session.last_todo_query.expect("missing todo snapshot");
-    assert_eq!(snapshot.query_type, "list");
-    assert_eq!(
-        snapshot.result_ids,
-        vec![first.id.clone(), second.id.clone()]
-    );
-
-    let _ = service
-        .respond(private_message("完成第一条"))
-        .await
-        .unwrap();
-    assert!(inspector.tool_call_count() >= 1);
-    let mut tool_requests = inspector.tool_requests();
-    let tool_request = tool_requests.pop().expect("missing tool request");
-    let completed_output = tool_request
-        .tools
-        .execute_json(
-            &tool_request.tool_context,
-            "complete_todos",
-            r#"{"numbers":[1],"reference":null}"#,
-        )
-        .await
-        .unwrap();
-    let completed: Value = serde_json::from_str(&completed_output).unwrap();
-    assert_eq!(completed["ok"], true);
-    assert_eq!(completed["completed"][0]["visible_number"], 1);
-    assert_eq!(completed["completed"][0]["title"], "测试代办");
-
-    assert_eq!(
-        service
-            .todo_store
-            .get_by_id(&owner, &first.id)
-            .unwrap()
-            .unwrap()
-            .status,
-        TodoStatus::Completed
-    );
-    assert_eq!(
-        service
-            .todo_store
-            .get_by_id(&owner, &second.id)
-            .unwrap()
-            .unwrap()
-            .status,
-        TodoStatus::Pending
-    );
-}
-
-#[tokio::test]
-async fn deterministic_date_query_then_tool_loop_complete_first_uses_date_snapshot() {
-    let inspector = MockProvider::new().with_tool_protocol(ToolCallingProtocol::OpenAiResponses);
-    let service = test_service_with_provider_and_tool_calling(inspector.clone(), true);
-    let owner = TodoStore::owner(Some("u1"), "private:u1");
-    let today = crate::util::time_context::request_time_context()
-        .local_date()
-        .format("%Y-%m-%d")
-        .to_string();
-    let today_item = service
-        .todo_store
-        .create(
-            &owner,
-            TodoItemDraft {
-                title: "今天要完成".to_owned(),
-                detail: None,
-                raw_text: None,
-                due_date: Some(today.clone()),
-                due_at: None,
-                reminder_at: None,
-                time_precision: TodoTimePrecision::Date,
-                recurrence_kind: crate::runtime::todo::TodoRecurrenceKind::None,
-                recurrence_interval_days: 0,
-                recurrence_interval: 0,
-                recurrence_unit: crate::runtime::todo::TodoRecurrenceUnit::Day,
-            },
-        )
-        .unwrap();
-    let no_time = service
-        .todo_store
-        .create(
-            &owner,
-            TodoItemDraft {
-                title: "无时间待办".to_owned(),
-                detail: None,
-                raw_text: None,
-                due_date: None,
-                due_at: None,
-                reminder_at: None,
-                time_precision: TodoTimePrecision::None,
-                recurrence_kind: crate::runtime::todo::TodoRecurrenceKind::None,
-                recurrence_interval_days: 0,
-                recurrence_interval: 0,
-                recurrence_unit: crate::runtime::todo::TodoRecurrenceUnit::Day,
-            },
-        )
-        .unwrap();
-
-    let listed = service
-        .respond(private_message("查看今天待办"))
-        .await
-        .unwrap();
-    assert_eq!(listed.command.as_deref(), Some("todo_due_date"));
-    let listed_text = listed.text.unwrap();
-    assert!(listed_text.contains("1. 今天要完成"));
-    assert!(!listed_text.contains("无时间待办"));
-
-    let session = service
-        .session_store
-        .get_or_create_active(&private_test_meta())
-        .unwrap();
-    let snapshot = session.last_todo_query.expect("missing date snapshot");
-    assert_eq!(snapshot.query_type, "due-date");
-    assert_eq!(snapshot.condition, today);
-    assert_eq!(snapshot.result_ids, vec![today_item.id.clone()]);
-
-    let _ = service
-        .respond(private_message("完成第一条"))
-        .await
-        .unwrap();
-    let mut tool_requests = inspector.tool_requests();
-    let tool_request = tool_requests.pop().expect("missing tool request");
-    let completed_output = tool_request
-        .tools
-        .execute_json(
-            &tool_request.tool_context,
-            "complete_todos",
-            r#"{"numbers":[1],"reference":null}"#,
-        )
-        .await
-        .unwrap();
-    let completed: Value = serde_json::from_str(&completed_output).unwrap();
-    assert_eq!(completed["ok"], true);
-    assert_eq!(completed["completed"][0]["title"], "今天要完成");
-
-    assert_eq!(
-        service
-            .todo_store
-            .get_by_id(&owner, &today_item.id)
-            .unwrap()
-            .unwrap()
-            .status,
-        TodoStatus::Completed
-    );
-    assert_eq!(
-        service
-            .todo_store
-            .get_by_id(&owner, &no_time.id)
-            .unwrap()
-            .unwrap()
-            .status,
-        TodoStatus::Pending
-    );
-}
-
-#[tokio::test]
-async fn deterministic_todo_query_alias_then_tool_loop_complete_first_uses_latest_snapshot() {
-    let inspector = MockProvider::new().with_tool_protocol(ToolCallingProtocol::OpenAiResponses);
-    let service = test_service_with_provider_and_tool_calling(inspector.clone(), true);
-    let owner = TodoStore::owner(Some("u1"), "private:u1");
-    let first = service
-        .todo_store
-        .create(
-            &owner,
-            TodoItemDraft {
-                title: "代办 A".to_owned(),
-                detail: None,
-                raw_text: None,
-                due_date: None,
-                due_at: None,
-                reminder_at: None,
-                time_precision: TodoTimePrecision::None,
-                recurrence_kind: crate::runtime::todo::TodoRecurrenceKind::None,
-                recurrence_interval_days: 0,
-                recurrence_interval: 0,
-                recurrence_unit: crate::runtime::todo::TodoRecurrenceUnit::Day,
-            },
-        )
-        .unwrap();
-    service
-        .todo_store
-        .create(
-            &owner,
-            TodoItemDraft {
-                title: "代办 B".to_owned(),
-                detail: None,
-                raw_text: None,
-                due_date: None,
-                due_at: None,
-                reminder_at: None,
-                time_precision: TodoTimePrecision::None,
-                recurrence_kind: crate::runtime::todo::TodoRecurrenceKind::None,
-                recurrence_interval_days: 0,
-                recurrence_interval: 0,
-                recurrence_unit: crate::runtime::todo::TodoRecurrenceUnit::Day,
-            },
-        )
-        .unwrap();
-
-    let listed = service
-        .respond(private_message("看一下代办"))
-        .await
-        .unwrap();
-    assert_eq!(listed.command.as_deref(), Some("todo_list"));
-    assert!(listed.text.as_deref().unwrap().contains("1. 代办 A"));
-
-    let _ = service
-        .respond(private_message("完成第一条"))
-        .await
-        .unwrap();
-    let tool_request = inspector
-        .tool_requests()
-        .pop()
-        .expect("missing tool request after alias query");
-    let completed_output = tool_request
-        .tools
-        .execute_json(
-            &tool_request.tool_context,
-            "complete_todos",
-            r#"{"numbers":[1],"reference":null}"#,
-        )
-        .await
-        .unwrap();
-    let completed: Value = serde_json::from_str(&completed_output).unwrap();
-    assert_eq!(completed["ok"], true);
-    assert_eq!(completed["completed"][0]["title"], "代办 A");
-    assert_eq!(
-        service
-            .todo_store
-            .get_by_id(&owner, &first.id)
-            .unwrap()
-            .unwrap()
-            .status,
-        TodoStatus::Completed
-    );
-}
-
-#[tokio::test]
-async fn todo_complete_receipt_refreshes_pending_list_and_collapses_one_hidden_item() {
-    let inspector = MockProvider::new()
-        .with_tool_protocol(ToolCallingProtocol::OpenAiResponses)
-        .with_tool_call_json(
-            "complete_todos",
-            r#"{"numbers":[1],"reference":null}"#,
-            "已完成第一条",
-        );
-    let service = test_service_with_provider_and_tool_calling(inspector.clone(), true);
-    let owner = TodoStore::owner(Some("u1"), "private:u1");
-    for index in 1..=7 {
-        service
-            .todo_store
-            .create(
-                &owner,
-                TodoItemDraft {
-                    title: format!("待办 {index}"),
-                    detail: None,
-                    raw_text: None,
-                    due_date: None,
-                    due_at: None,
-                    reminder_at: None,
-                    time_precision: TodoTimePrecision::None,
-                    recurrence_kind: crate::runtime::todo::TodoRecurrenceKind::None,
-                    recurrence_interval_days: 0,
-                    recurrence_interval: 0,
-                    recurrence_unit: crate::runtime::todo::TodoRecurrenceUnit::Day,
-                },
-            )
-            .unwrap();
-    }
-
-    service
-        .respond(private_message("看一下待办"))
-        .await
-        .unwrap();
-    let response = service
-        .respond(private_message("完成第一条"))
-        .await
-        .unwrap();
-
-    let text = response.text.unwrap();
-    assert!(text.contains("✅ 已完成待办 · 1条"));
-    assert!(text.contains("🚧 当前进行中 · 共 6 项"));
-    assert!(text.contains("还有 1 项进行中待办，可说“查看全部进行中待办”。"));
-    let remaining = service.todo_store.list_pending(&owner).unwrap();
-    let session = service
-        .session_store
-        .get_or_create_active(&private_test_meta())
-        .unwrap();
-    let snapshot = session.last_todo_query.expect("missing refreshed snapshot");
-    assert_eq!(snapshot.query_type, "list");
-    assert_eq!(
-        snapshot.result_ids,
-        remaining
-            .iter()
-            .take(5)
-            .map(|item| item.id.clone())
-            .collect::<Vec<_>>()
-    );
-    assert_eq!(snapshot.result_ids.len(), 5);
-}
-
-#[tokio::test]
-async fn todo_complete_receipt_collapses_two_hidden_items() {
-    let inspector = MockProvider::new()
-        .with_tool_protocol(ToolCallingProtocol::OpenAiResponses)
-        .with_tool_call_json(
-            "complete_todos",
-            r#"{"numbers":[1],"reference":null}"#,
-            "已完成第一条",
-        );
-    let service = test_service_with_provider_and_tool_calling(inspector.clone(), true);
-    let owner = TodoStore::owner(Some("u1"), "private:u1");
-    for index in 1..=8 {
-        service
-            .todo_store
-            .create(&owner, todo_draft(format!("待办 {index}")))
-            .unwrap();
-    }
-
-    service
-        .respond(private_message("看一下待办"))
-        .await
-        .unwrap();
-    let response = service
-        .respond(private_message("完成第一条"))
-        .await
-        .unwrap();
-
-    let text = response.text.unwrap();
-    assert!(text.contains("🚧 当前进行中 · 共 7 项"));
-    assert!(text.contains("还有 2 项进行中待办，可说“查看全部进行中待办”。"));
-    let remaining = service.todo_store.list_pending(&owner).unwrap();
-    let session = service
-        .session_store
-        .get_or_create_active(&private_test_meta())
-        .unwrap();
-    let snapshot = session.last_todo_query.expect("missing refreshed snapshot");
-    assert_eq!(
-        snapshot.result_ids,
-        remaining
-            .iter()
-            .take(5)
-            .map(|item| item.id.clone())
-            .collect::<Vec<_>>()
-    );
-    assert_eq!(snapshot.result_ids.len(), 5);
-}
-
-#[tokio::test]
-async fn todo_complete_receipt_uses_dynamic_pending_collapse_hint() {
-    let inspector = MockProvider::new()
-        .with_tool_protocol(ToolCallingProtocol::OpenAiResponses)
-        .with_tool_call_json(
-            "complete_todos",
-            r#"{"numbers":[1],"reference":null}"#,
-            "已完成第一条",
-        );
-    let service = test_service_with_provider_and_tool_calling(inspector.clone(), true);
-    let owner = TodoStore::owner(Some("u1"), "private:u1");
-    for index in 1..=9 {
-        service
-            .todo_store
-            .create(&owner, todo_draft(format!("待办 {index}")))
-            .unwrap();
-    }
-
-    service
-        .respond(private_message("看一下待办"))
-        .await
-        .unwrap();
-    let response = service
-        .respond(private_message("完成第一条"))
-        .await
-        .unwrap();
-
-    let text = response.text.unwrap();
-    assert!(text.contains("🚧 当前进行中 · 共 8 项"));
-    assert!(text.contains("还有 3 项进行中待办，可说“查看全部进行中待办”。"));
-    let remaining = service.todo_store.list_pending(&owner).unwrap();
-    let session = service
-        .session_store
-        .get_or_create_active(&private_test_meta())
-        .unwrap();
-    let snapshot = session.last_todo_query.expect("missing refreshed snapshot");
-    assert_eq!(snapshot.query_type, "list");
-    assert_eq!(
-        snapshot.result_ids,
-        remaining
-            .iter()
-            .take(5)
-            .map(|item| item.id.clone())
-            .collect::<Vec<_>>()
-    );
-    assert_eq!(snapshot.result_ids.len(), 5);
-}
-
-#[tokio::test]
-async fn deterministic_completed_query_then_tool_loop_restore_first_uses_latest_snapshot() {
-    let inspector = MockProvider::new().with_tool_protocol(ToolCallingProtocol::OpenAiResponses);
-    let service = test_service_with_provider_and_tool_calling(inspector.clone(), true);
-    let owner = TodoStore::owner(Some("u1"), "private:u1");
-    let first = service
-        .todo_store
-        .create(
-            &owner,
-            TodoItemDraft {
-                title: "已完成 A".to_owned(),
-                detail: None,
-                raw_text: None,
-                due_date: None,
-                due_at: None,
-                reminder_at: None,
-                time_precision: TodoTimePrecision::None,
-                recurrence_kind: crate::runtime::todo::TodoRecurrenceKind::None,
-                recurrence_interval_days: 0,
-                recurrence_interval: 0,
-                recurrence_unit: crate::runtime::todo::TodoRecurrenceUnit::Day,
-            },
-        )
-        .unwrap();
-    let second = service
-        .todo_store
-        .create(
-            &owner,
-            TodoItemDraft {
-                title: "已完成 B".to_owned(),
-                detail: None,
-                raw_text: None,
-                due_date: None,
-                due_at: None,
-                reminder_at: None,
-                time_precision: TodoTimePrecision::None,
-                recurrence_kind: crate::runtime::todo::TodoRecurrenceKind::None,
-                recurrence_interval_days: 0,
-                recurrence_interval: 0,
-                recurrence_unit: crate::runtime::todo::TodoRecurrenceUnit::Day,
-            },
-        )
-        .unwrap();
-    service.todo_store.complete(&owner, &first.id).unwrap();
-    service.todo_store.complete(&owner, &second.id).unwrap();
-
-    let listed = service
-        .respond(private_message("看看已完成"))
-        .await
-        .unwrap();
-    assert_eq!(listed.command.as_deref(), Some("todo_done"));
-    let session = service
-        .session_store
-        .get_or_create_active(&private_test_meta())
-        .unwrap();
-    let snapshot = session.last_todo_query.expect("missing completed snapshot");
-    assert_eq!(snapshot.query_type, "completed-list");
-    let expected_first_id = snapshot
-        .result_ids
-        .first()
-        .cloned()
-        .expect("completed snapshot should contain first item");
-    let expected_first_title = service
-        .todo_store
-        .get_by_id(&owner, &expected_first_id)
-        .unwrap()
-        .unwrap()
-        .title;
-
-    let _ = service
-        .respond(private_message("恢复第一条"))
-        .await
-        .unwrap();
-    let tool_request = inspector
-        .tool_requests()
-        .pop()
-        .expect("missing restore tool request");
-    let restored_output = tool_request
-        .tools
-        .execute_json(
-            &tool_request.tool_context,
-            "restore_todos",
-            r#"{"numbers":[1],"reference":null}"#,
-        )
-        .await
-        .unwrap();
-    let restored: Value = serde_json::from_str(&restored_output).unwrap();
-    assert_eq!(restored["ok"], true);
-    assert_eq!(restored["restored"][0]["title"], expected_first_title);
-    assert_eq!(
-        service
-            .todo_store
-            .get_by_id(&owner, &expected_first_id)
-            .unwrap()
-            .unwrap()
-            .status,
-        TodoStatus::Pending
-    );
-}
-
-#[tokio::test]
-async fn deterministic_cancelled_query_then_tool_loop_restore_first_uses_latest_snapshot() {
-    let inspector = MockProvider::new().with_tool_protocol(ToolCallingProtocol::OpenAiResponses);
-    let service = test_service_with_provider_and_tool_calling(inspector.clone(), true);
-    let owner = TodoStore::owner(Some("u1"), "private:u1");
-    let first = service
-        .todo_store
-        .create(
-            &owner,
-            TodoItemDraft {
-                title: "已取消 A".to_owned(),
-                detail: None,
-                raw_text: None,
-                due_date: None,
-                due_at: None,
-                reminder_at: None,
-                time_precision: TodoTimePrecision::None,
-                recurrence_kind: crate::runtime::todo::TodoRecurrenceKind::None,
-                recurrence_interval_days: 0,
-                recurrence_interval: 0,
-                recurrence_unit: crate::runtime::todo::TodoRecurrenceUnit::Day,
-            },
-        )
-        .unwrap();
-    let second = service
-        .todo_store
-        .create(
-            &owner,
-            TodoItemDraft {
-                title: "已取消 B".to_owned(),
-                detail: None,
-                raw_text: None,
-                due_date: None,
-                due_at: None,
-                reminder_at: None,
-                time_precision: TodoTimePrecision::None,
-                recurrence_kind: crate::runtime::todo::TodoRecurrenceKind::None,
-                recurrence_interval_days: 0,
-                recurrence_interval: 0,
-                recurrence_unit: crate::runtime::todo::TodoRecurrenceUnit::Day,
-            },
-        )
-        .unwrap();
-    service.todo_store.cancel(&owner, &first.id).unwrap();
-    service.todo_store.cancel(&owner, &second.id).unwrap();
-
-    let listed = service
-        .respond(private_message("看看已取消"))
-        .await
-        .unwrap();
-    assert_eq!(listed.command.as_deref(), Some("todo_cancelled_list"));
-    let session = service
-        .session_store
-        .get_or_create_active(&private_test_meta())
-        .unwrap();
-    let snapshot = session.last_todo_query.expect("missing cancelled snapshot");
-    let expected_first_id = snapshot
-        .result_ids
-        .first()
-        .cloned()
-        .expect("cancelled snapshot should contain first item");
-    let expected_first_title = service
-        .todo_store
-        .get_by_id(&owner, &expected_first_id)
-        .unwrap()
-        .unwrap()
-        .title;
-
-    let _ = service
-        .respond(private_message("恢复第一条"))
-        .await
-        .unwrap();
-    let tool_request = inspector
-        .tool_requests()
-        .pop()
-        .expect("missing cancelled restore tool request");
-    let restored_output = tool_request
-        .tools
-        .execute_json(
-            &tool_request.tool_context,
-            "restore_todos",
-            r#"{"numbers":[1],"reference":null}"#,
-        )
-        .await
-        .unwrap();
-    let restored: Value = serde_json::from_str(&restored_output).unwrap();
-    assert_eq!(restored["ok"], true);
-    assert_eq!(restored["restored"][0]["title"], expected_first_title);
-    assert_eq!(
-        service
-            .todo_store
-            .get_by_id(&owner, &expected_first_id)
-            .unwrap()
-            .unwrap()
-            .status,
-        TodoStatus::Pending
-    );
-}
-
-#[tokio::test]
 async fn cancelled_query_then_delete_all_creates_bulk_pending_and_confirm_deletes_only_cancelled() {
     let inspector = MockProvider::new()
         .with_tool_protocol(ToolCallingProtocol::OpenAiResponses)
@@ -4205,133 +3355,6 @@ async fn cancelled_query_then_delete_all_creates_bulk_pending_and_confirm_delete
     let pending = service.todo_store.list_pending(&owner).unwrap();
     assert_eq!(pending.len(), 1);
     assert_eq!(pending[0].id, pending_a.id);
-}
-
-#[tokio::test]
-async fn deterministic_empty_query_clears_old_snapshot_before_number_mutation() {
-    let inspector = MockProvider::new().with_tool_protocol(ToolCallingProtocol::OpenAiResponses);
-    let service = test_service_with_provider_and_tool_calling(inspector.clone(), true);
-    let owner = TodoStore::owner(Some("u1"), "private:u1");
-    let todo = service
-        .todo_store
-        .create(
-            &owner,
-            TodoItemDraft {
-                title: "旧快照条目".to_owned(),
-                detail: None,
-                raw_text: None,
-                due_date: None,
-                due_at: None,
-                reminder_at: None,
-                time_precision: TodoTimePrecision::None,
-                recurrence_kind: crate::runtime::todo::TodoRecurrenceKind::None,
-                recurrence_interval_days: 0,
-                recurrence_interval: 0,
-                recurrence_unit: crate::runtime::todo::TodoRecurrenceUnit::Day,
-            },
-        )
-        .unwrap();
-
-    service
-        .respond(private_message("看一下待办"))
-        .await
-        .unwrap();
-    service.todo_store.complete(&owner, &todo.id).unwrap();
-
-    let empty_list = service
-        .respond(private_message("看一下待办"))
-        .await
-        .unwrap();
-    assert!(
-        empty_list
-            .text
-            .as_deref()
-            .unwrap()
-            .contains("暂无未完成待办")
-    );
-    let session = service
-        .session_store
-        .get_or_create_active(&private_test_meta())
-        .unwrap();
-    let snapshot = session.last_todo_query.expect("missing empty snapshot");
-    assert!(snapshot.result_ids.is_empty());
-
-    let _ = service
-        .respond(private_message("完成第一条"))
-        .await
-        .unwrap();
-    let tool_request = inspector
-        .tool_requests()
-        .pop()
-        .expect("missing tool request after empty query");
-    let completed_output = tool_request
-        .tools
-        .execute_json(
-            &tool_request.tool_context,
-            "complete_todos",
-            r#"{"numbers":[1],"reference":null}"#,
-        )
-        .await
-        .unwrap();
-    let completed: Value = serde_json::from_str(&completed_output).unwrap();
-    assert_eq!(completed["ok"], false);
-    assert_eq!(completed["requires_clarification"], true);
-    assert_eq!(completed["pending_action"], "clarify");
-}
-
-#[tokio::test]
-async fn deterministic_query_then_status_changes_returns_precise_missing_error() {
-    let inspector = MockProvider::new().with_tool_protocol(ToolCallingProtocol::OpenAiResponses);
-    let service = test_service_with_provider_and_tool_calling(inspector.clone(), true);
-    let owner = TodoStore::owner(Some("u1"), "private:u1");
-    let todo = service
-        .todo_store
-        .create(
-            &owner,
-            TodoItemDraft {
-                title: "状态先被改掉".to_owned(),
-                detail: None,
-                raw_text: None,
-                due_date: None,
-                due_at: None,
-                reminder_at: None,
-                time_precision: TodoTimePrecision::None,
-                recurrence_kind: crate::runtime::todo::TodoRecurrenceKind::None,
-                recurrence_interval_days: 0,
-                recurrence_interval: 0,
-                recurrence_unit: crate::runtime::todo::TodoRecurrenceUnit::Day,
-            },
-        )
-        .unwrap();
-
-    service
-        .respond(private_message("看一下待办"))
-        .await
-        .unwrap();
-    // 模拟用户看到列表后，条目已被其他操作提前完成。
-    service.todo_store.complete(&owner, &todo.id).unwrap();
-
-    let _ = service
-        .respond(private_message("完成第一条"))
-        .await
-        .unwrap();
-    let tool_request = inspector
-        .tool_requests()
-        .pop()
-        .expect("missing tool request after state change");
-    let completed_output = tool_request
-        .tools
-        .execute_json(
-            &tool_request.tool_context,
-            "complete_todos",
-            r#"{"numbers":[1],"reference":null}"#,
-        )
-        .await
-        .unwrap();
-    let completed: Value = serde_json::from_str(&completed_output).unwrap();
-    assert_eq!(completed["ok"], true);
-    assert_eq!(completed["completed"], serde_json::json!([]));
-    assert_eq!(completed["missing_numbers"], serde_json::json!([1]));
 }
 
 #[tokio::test]
