+++ qq-maid-core/src/runtime/respond/tests/mod.rs
@@ -7,6 +7,8 @@ mod search;
 mod session;
 mod support;
 mod todo;
+mod todo_receipt;
+mod todo_tool_loop;
 mod train;
 mod translation;
 mod weather;
